Reordering methods to be compliant with StyleCop analyzers
